separated pages and components
