+++ front/src/app/pages/home/home.component.html
