+++ front/src/app/pages/home/home.component.scss
